+++ editor/debugger/editor_debugger_inspector.cpp
@@ -167,7 +167,9 @@ EditorDebuggerRemoteObjects *EditorDebuggerInspector::set_objects(const Array &p
 		SceneDebuggerObject obj;
 		obj.deserialize(arr);
 		if (obj.id.is_valid()) {
-			ids.push_back((uint64_t)obj.id);
+			// The element type of a `TypedArray<uint64_t>` is declared `Variant::INT`, so the
+			// instance id is carried nominally rather than in the unsigned carrier.
+			ids.push_back((int64_t)(uint64_t)obj.id);
 			objects.push_back(obj);
 		}
 	}
